refactor: split orders near-cap lib modules and ratchet ceiling to 300 (#4508)
Extract explorer prospect, diplomatic targets, army-move destinations, completed-work special handlers, validator factory wiring, and spy/regiment helpers so every orders lib file stays under the wave-9 300-line ceiling.
