+++ packages/colonizethis_orders/lib/src/orders/orders_application_completed_work.dart
@@ -3,6 +3,7 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'order_work_constants.dart';
 import 'orders_application_context.dart';
 import 'orders_application_completed_work_handlers.dart';
+import 'orders_application_completed_work_handlers_special.dart';
 import 'orders_application_road_propagation.dart';
 
 export 'orders_application_completed_work_handlers.dart'
